# E3_EOS_Complete_MD_2026-09-26/contracts/CORE_COMMANDS.openapi.yaml
openapi: 3.1.0
info:
  title: E3-EOS Core Command Contracts
  version: 1.0.0
  description: Normative starting subset for consequential commands, not a complete
    generated implementation. Universal policy/current-authority/idempotency rules
    apply. Extend remaining CRUD and observation-specific schemas per phase.
servers:
- url: /api/v1
security:
- eosSession: []
paths:
...
  /projects/{projectId}/reports/{id}/publish:
    post:
      operationId: publishReport
      summary: publishReport
      x-eos-permission: report.publish
      parameters:
      - name: projectId
        in: path
        required: true
        schema: *id001
      - name: id
        in: path
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 128
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
          minLength: 1
        description: Current ETag of addressed aggregate/draft, in addition to target
          content version/hash.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportPublish'
      responses:
        '201':
          description: Recorded result or queued durable command; never implies unverified
            external completion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          description: Malformed request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Unauthenticated
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden action
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found or hidden scope
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Conflicting state/idempotency/reservation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '412':
          description: Stale record precondition
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation, missing condition or authority
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '428':
          description: Missing required precondition
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Rate limited
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Current required authority/source unavailable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /projects/{projectId}/closure-decisions:
    post:
      operationId: closeProjectDimension
      summary: closeProjectDimension
      x-eos-permission: project.close
      parameters:
      - name: projectId
        in: path
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseDimension'
      responses:
        '201':
          description: Recorded result or queued durable command; never implies unverified
            external completion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          description: Malformed request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Unauthenticated
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden action
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found or hidden scope
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Conflicting state/idempotency/reservation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '412':
          description: Stale record precondition
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation, missing condition or authority
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '428':
          description: Missing required precondition
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Rate limited
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: Current required authority/source unavailable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    eosSession:
      type: apiKey
      in: cookie
      name: __Host-eos-session
      description: Logical session-cookie contract. Map the tested Better Auth cookie
        configuration to this name or update this contract before implementation;
        server-side sessions and CSRF/origin enforcement required.
  schemas:
...
    ReportPublish:
      type: object
      additionalProperties: false
      properties:
        reportVersionId: *id001
        sourceSnapshotId: *id001
        targetHash: *id003
        audience:
          type: string
          enum:
          - internal
          - client
          - restricted
        approvalDecisionIds:
          type: array
          items: *id001
      required:
      - reportVersionId
      - sourceSnapshotId
      - targetHash
      - audience
      - approvalDecisionIds
    CloseDimension:
      type: object
      additionalProperties: false
      properties:
        dimension:
          type: string
          enum:
          - operational
          - acceptance
          - reporting
          - financial_review
          - settlement
        evidenceManifestId: *id001
        decisionIds:
          type: array
          items: *id001
        disclosedOpenItems:
          type: array
          items: *id001
      required:
      - dimension
      - evidenceManifestId
      - decisionIds
      - disclosedOpenItems
...
    CommandResult:
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: false
          properties:
            id: *id001
            status:
              type: string
            recordVersion:
              type: integer
              minimum: 1
            jobId: *id001
            externalDeliveryStatus:
              type: string
              enum:
              - not_applicable
              - queued
              - delivered
              - acknowledged
              - ambiguous
              - failed
          required:
          - id
          - status
          - recordVersion
        meta:
          type: object
          additionalProperties: false
          properties:
            requestId:
              type: string
            policySnapshotId: *id001
            dataAsOf: *id002
          required:
          - requestId
      required:
      - data
      - meta
    Problem:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        requestId:
          type: string
        conditions:
          type: array
          items:
            type: object
            additionalProperties: true
        permittedNextActions:
          type: array
          items:
            type: string
      required:
      - type
      - title
      - status
      - code
      - requestId